finish layout overhaul
