--- gui/duaas.txt
@@ -1,17 +1,0 @@
- (x3)أَسْتَغْفِرُ اللّٰه
-||
-Astagfirullah (x3)
-||
-I seek forgiveness from Allah
----
- اللَّهُمَّ أَنْتَ السَّلاَمُ وَمِنْكَ السَّلاَمُ تَبَارَكْتَ يَا ذَا الْجَلاَلِ وَالإِكْرَامِ
-||
-Allahumma antas-salaam wa minkas-salaam. Tabarakta ya Dhal-jalali wal- ikram.
-||
-O Allah, You are the Flawless and the Source of Peace, from You is all peace, blessed are You O Possessor of Majesty and Honour
----
-اللَّهُمَّ أَعِنِّي عَلَى ذِكْرِكَ، وَشُكْرِكَ، وَحُسْنِ عِبَادَتِكَ
-||
-Allâhumma a’innî alâ dhikrika, wa shukrika, wa husni ‘ibâdatika
-||
-O Allah, help me remember You, to be grateful to You, and to worship You in an excellent manner
