--- gui/new_main.py
@@ -1,7 +1,0 @@
-from new_gui import GUI
-def main():
-    gui = GUI()
-    gui.run()
-
-if __name__ == '__main__':
-    main()
